feat: refactored-user-permissions-member-table
